backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
tensorflow==2.20.0
pillow==11.0.0
python-multipart==0.0.12
numpy==2.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
email-validator==2.2.0
